日常提交

+++ games101_01/src/triangle.rs
@@ -31,22 +31,22 @@ impl Triangle{
     }
 
     pub fn set_vertex(&mut self, ind:i32, ver:Vector3f){
-        self.v[ind] = ver;
+        self.v[ind as usize] = ver;
     }
 
     pub fn set_normal(&mut self, ind:i32, n:Vector3f){
-        self.normal[ind] = n;
+        self.normal[ind as usize] = n;
     }
 
     pub fn set_color(&mut self, ind:i32, r:f32, g:f32, b:f32){
         if r<0.0||r>255.0||g<0.0||g>255.0||b<0.0||b>255.0{
             panic!("Invalid color values");
         }
-        self.color[ind] = Vector3f::new(r/255f32,g/255f32,b/255f32);
+        self.color[ind as usize] = Vector3f::new(r/255f32,g/255f32,b/255f32);
     }
 
     pub fn set_tex_coord(&mut self, ind:i32, s:f32, t:f32){
-        self.tex_coords[ind] = Vector2f::new(s,t);
+        self.tex_coords[ind as usize] = Vector2f::new(s,t);
     }
 
     pub fn to_vector4(&self) ->[Vector4f;3]{
